renamed GridComponent
